fix install script

+++ bin/install.sh
@@ -37,9 +37,14 @@ if [ "$(id -u)" -ne 0 ]; then
     SUDO="sudo"
 fi
 
-JDK_MIRROR_URL="https://mirrors.tuna.tsinghua.edu.cn/Adoptium/21/jdk/${ARCH}/linux/OpenJDK21U-jdk_${ARCH}_linux_hotspot_21.0.4_7.tar.gz"
-status "Install JDK from ${JDK_MIRROR_URL}"
-curl --fail --show-error --location --progress-bar -o "/tmp/jdk21.tar.gz" ${JDK_MIRROR_URL}
+if [ ! -f /tmp/jdk21.tar.gz ]; then
+    JDK_MIRROR_URL="https://mirrors.tuna.tsinghua.edu.cn/Adoptium/21/jdk/${ARCH}/linux/OpenJDK21U-jdk_${ARCH}_linux_hotspot_21.0.4_7.tar.gz"
+    status "Installing JDK from ${JDK_MIRROR_URL}"
+    curl --fail --show-error --location --progress-bar -o "/tmp/jdk21.tar.gz" ${JDK_MIRROR_URL}
+else
+    status "JDK exists hence no need to download"
+fi
+
 mkdir -p /tmp/jdk21
 tar -zxf /tmp/jdk21.tar.gz --strip-components 1 -C /tmp/jdk21
 export PATH=/tmp/jdk21/bin:$PATH
